skip negative id when remove
git-svn-id: https://svn.code.sf.net/p/moneymanagerex/code/trunk@5098 7f5d68d0-c712-0410-839f-9b116a576cc5

+++ util/sqlite2cpp.py
@@ -364,6 +364,7 @@ struct DB_Table_%s : public DB_Table
     /** Remove the Data record from the database and the memory table (cashe)*/
     bool remove(int id, wxSQLite3Database* db)
     {
+        if (id < 0) return false;
         try
         {
             wxString sql = "DELETE FROM %s WHERE %s = ?";
